[win10] prevent display switch off while playing media.
