+++ xbmc/powermanagement/CMakeLists.txt
@@ -6,4 +6,8 @@ set(HEADERS DPMSSupport.h
             IPowerSyscall.h
             PowerManager.h)
 
+if(CORE_SYSTEM_NAME MATCHES windows)
+  list(APPEND HEADERS WinIdleTimer.h)
+endif()
+
 core_add_library(powermanagement)
